requirements.txt
absl-py==2.1.0
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrgh==1.0.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
asyncssh==2.17.0
attrs==24.2.0
Babel==2.10.3
backports.shutil-get-terminal-size==1.0.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
brotlipy==0.7.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.5.18.1
cffi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/croot-jgj0vmyy/cffi_1642701117808/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.86
citeproc-py==0.6.0
cleo==2.0.1
click==8.1.3
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_croot-4sc96bd_/conda-package-handling_1649105290173/work
contourpy==1.0.7
crashtest==0.4.1
cryptography==40.0.2
cycler==0.11.0
dacite==1.8.1
debugpy==1.8.5
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
distrax==0.1.5
dm-tree==0.1.8
docker-pycreds==0.4.0
docutils==0.19
duecredit==0.10.2
dulwich==0.21.3
etils==1.5.2
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.16.1
ffmpeg==1.4
fica==0.2.0
filelock==3.9.0
flax==0.8.5
fonttools==4.38.0
fqdn==1.5.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.10.0
google-api-python-client==2.60.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
graphviz==0.20.3
gspread==5.5.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.4
httpx==0.27.0
-e git+ssh://git@github.com/spike-imperial/ued-hrm.git@4d8e3e695cd62e4172330baf62e58a995c1993bb#egg=hurms
hydra-core==1.3.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.35.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.12.0
iniconfig==2.0.0
installer==0.6.0
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.3
isoduration==20.11.0
jaraco.classes==3.2.3
jax==0.4.30
jax-tqdm==0.2.2
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
jupytext==1.14.1
keyring==23.13.1
kiwisolver==1.4.4
lockfile==0.12.2
looseversion==1.3.0
lxml==4.9.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.0
mdurl==0.1.2
meinsweeper==0.1.4
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==9.1.0
moviepy==1.0.3
msgpack==1.0.4
myslideslive==0.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.7
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.1
omegaconf==2.3.0
opencv-python==4.10.0.84
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.23
otter-grader==4.0.1
overrides==7.7.0
packaging==24.1
pandas==1.4.4
pandocfilters==1.5.1
paramiko==3.4.1
parso==0.8.4
pexpect==4.8.0
Pillow==9.4.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.5.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.21.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.2
Pygments==2.13.0
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.18.1
PySocks @ file:///Users/ktietz/Code/oss/ci_pkgs/pysocks_1626781349491/work
pytest==8.3.2
pytest-subtests==0.13.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-on-whales==0.52.0
pytz==2022.2.1
PyYAML==6.0
pyzmq==26.1.0
qbstyles==0.1.4
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==2.13.7
referencing==0.35.1
reprint==0.6.0
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.20.0
rsa==4.9
ruamel-yaml-conda @ file:///Users/ktietz/demo/mc3/conda-bld/ruamel_yaml_1629464769899/work
scipy==1.13.1
Send2Trash==1.8.3
sentry-sdk==2.13.0
setproctitle==1.3.3
shellingham==1.5.0.post1
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.6.3
tensorflow-probability==0.24.0
tensorstore==0.1.64
terminado==0.18.1
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trove-classifiers==2023.2.20
typer==0.6.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.3.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.20.0
wandb==0.17.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
wrapt==1.14.1
xattr==0.10.1
xminigrid==0.9.0
zipp==3.8.1
